feat: Igualar fluxo EFD Fiscal Simples ao Completo (v1.0.1)
- Implementar detecção de resposta assíncrona no SpedUploadService
  para modalidade gratuita quando webhook retorna apenas message
- Adicionar tratamento de data.async=true no frontend para conectar SSE
- Corrigir SSE para não retornar relatórios antigos usando timestamp
  de conexão (connectionEstablishedAt)
- Documentar fluxo completo de ambas modalidades (gratuita e completa)
- Versão 1.0.1 com todas as correções

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->unsignedInteger('credits')->default(0);
+            $table->decimal('credits', 12, 2)->default(0);
             $table->timestamps();
         });
